create P2PSyncer class for checkpoints, set checkpoint interval daily, remove checkpoint threading
